+++ runtime/__init__.py
@@ -3,12 +3,13 @@
 from .skill_loader import SkillLoader
 from .backends import LocalBackend
 from .evolution_gate import EvolutionCandidate, EvaluationResult, EvolutionGate
-from .skill_memory import SkillMemoryManager
+from .skill_memory import LearningSignal, SkillMemoryManager
 
 __all__ = [
     "LocalBackend",
     "SkillLoader",
     "SkillMemoryManager",
+    "LearningSignal",
     "EvolutionCandidate",
     "EvaluationResult",
     "EvolutionGate",